Reduce usage of InternalThreadLocalMap (#11999)
Motivation:

We should not use InternalThreadLocalMap whenever possible and just use FastThreadLocal.
This allows us to not expose InternalThreadLocalMap as a public class at all.

Modifications:

Replace usage of InternalThreadLocalMap with FastThreadLocal when possible.
Move InternalThreadLLocalMap to concurrent package and make it package-private.

Result:

Cleaner code and cleaner public API.
