+++ common/src/main/java/io/netty/util/concurrent/FastThreadLocal.java
@@ -15,7 +15,6 @@
  */
 package io.netty.util.concurrent;
 
-import io.netty.util.internal.InternalThreadLocalMap;
 import io.netty.util.internal.PlatformDependent;
 
 import java.util.Collections;
@@ -155,20 +154,6 @@ public class FastThreadLocal<V> {
             }
         }
         return null;
-    }
-
-    /**
-     * Returns the current value for the specified thread local map.
-     * The specified thread local map must be for the current thread.
-     */
-    @SuppressWarnings("unchecked")
-    public final V get(InternalThreadLocalMap threadLocalMap) {
-        Object v = threadLocalMap.indexedVariable(index);
-        if (v != InternalThreadLocalMap.UNSET) {
-            return (V) v;
-        }
-
-        return initialize(threadLocalMap);
     }
 
     private V initialize(InternalThreadLocalMap threadLocalMap) {
