+++ common/src/main/java/io/netty/util/concurrent/InternalThreadLocalMap.java
@@ -1,0 +1,197 @@
+/*
+ * Copyright 2014 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.netty.util.concurrent;
+
+import java.util.Arrays;
+import java.util.BitSet;
+import java.util.concurrent.atomic.AtomicInteger;
+
+/**
+ * The internal data structure that stores the thread-local variables for Netty and all {@link FastThreadLocal}s.
+ * Note that this class is for internal use only and is subject to change at any time.  Use {@link FastThreadLocal}
+ * unless you know what you are doing.
+ */
+final class InternalThreadLocalMap {
+
+    private static final ThreadLocal<InternalThreadLocalMap> slowThreadLocalMap =
+            new ThreadLocal<>();
+    private static final AtomicInteger nextIndex = new AtomicInteger();
+
+    private static final int ARRAY_LIST_CAPACITY_EXPAND_THRESHOLD = 1 << 30;
+    // Reference: https://hg.openjdk.java.net/jdk8/jdk8/jdk/file/tip/src/share/classes/java/util/ArrayList.java#l229
+    private static final int ARRAY_LIST_CAPACITY_MAX_SIZE = Integer.MAX_VALUE - 8;
+    private static final int INDEXED_VARIABLE_TABLE_INITIAL_SIZE = 32;
+
+    static final Object UNSET = new Object();
+
+    /** Used by {@link FastThreadLocal} */
+    private Object[] indexedVariables;
+
+    private BitSet cleanerFlags;
+
+    static InternalThreadLocalMap getIfSet() {
+        Thread thread = Thread.currentThread();
+        if (thread instanceof FastThreadLocalThread) {
+            return ((FastThreadLocalThread) thread).threadLocalMap();
+        }
+        return slowThreadLocalMap.get();
+    }
+
+    static InternalThreadLocalMap get() {
+        Thread thread = Thread.currentThread();
+        if (thread instanceof FastThreadLocalThread) {
+            return fastGet((FastThreadLocalThread) thread);
+        } else {
+            return slowGet();
+        }
+    }
+
+    private static InternalThreadLocalMap fastGet(FastThreadLocalThread thread) {
+        InternalThreadLocalMap threadLocalMap = thread.threadLocalMap();
+        if (threadLocalMap == null) {
+            thread.setThreadLocalMap(threadLocalMap = new InternalThreadLocalMap());
+        }
+        return threadLocalMap;
+    }
+
+    private static InternalThreadLocalMap slowGet() {
+        InternalThreadLocalMap ret = slowThreadLocalMap.get();
+        if (ret == null) {
+            ret = new InternalThreadLocalMap();
+            slowThreadLocalMap.set(ret);
+        }
+        return ret;
+    }
+
+    static void remove() {
+        Thread thread = Thread.currentThread();
+        if (thread instanceof FastThreadLocalThread) {
+            ((FastThreadLocalThread) thread).setThreadLocalMap(null);
+        } else {
+            slowThreadLocalMap.remove();
+        }
+    }
+
+    static void destroy() {
+        slowThreadLocalMap.remove();
+    }
+
+    static int nextVariableIndex() {
+        int index = nextIndex.getAndIncrement();
+        if (index >= ARRAY_LIST_CAPACITY_MAX_SIZE || index < 0) {
+            nextIndex.set(ARRAY_LIST_CAPACITY_MAX_SIZE);
+            throw new IllegalStateException("too many thread-local indexed variables");
+        }
+        return index;
+    }
+
+    static int lastVariableIndex() {
+        return nextIndex.get() - 1;
+    }
+
+    private InternalThreadLocalMap() {
+        indexedVariables = newIndexedVariableTable();
+    }
+
+    private static Object[] newIndexedVariableTable() {
+        Object[] array = new Object[INDEXED_VARIABLE_TABLE_INITIAL_SIZE];
+        Arrays.fill(array, UNSET);
+        return array;
+    }
+
+    int size() {
+        int count = 0;
+        for (Object o: indexedVariables) {
+            if (o != UNSET) {
+                count ++;
+            }
+        }
+
+        // We should subtract 1 from the count because the first element in 'indexedVariables' is reserved
+        // by 'FastThreadLocal' to keep the list of 'FastThreadLocal's to remove on 'FastThreadLocal.removeAll()'.
+        return count - 1;
+    }
+
+    Object indexedVariable(int index) {
+        Object[] lookup = indexedVariables;
+        return index < lookup.length? lookup[index] : UNSET;
+    }
+
+    /**
+     * @return {@code true} if and only if a new thread-local variable has been created
+     */
+    boolean setIndexedVariable(int index, Object value) {
+        Object[] lookup = indexedVariables;
+        if (index < lookup.length) {
+            Object oldValue = lookup[index];
+            lookup[index] = value;
+            return oldValue == UNSET;
+        } else {
+            expandIndexedVariableTableAndSet(index, value);
+            return true;
+        }
+    }
+
+    private void expandIndexedVariableTableAndSet(int index, Object value) {
+        Object[] oldArray = indexedVariables;
+        final int oldCapacity = oldArray.length;
+        int newCapacity;
+        if (index < ARRAY_LIST_CAPACITY_EXPAND_THRESHOLD) {
+            newCapacity = index;
+            newCapacity |= newCapacity >>>  1;
+            newCapacity |= newCapacity >>>  2;
+            newCapacity |= newCapacity >>>  4;
+            newCapacity |= newCapacity >>>  8;
+            newCapacity |= newCapacity >>> 16;
+            newCapacity ++;
+        } else {
+            newCapacity = ARRAY_LIST_CAPACITY_MAX_SIZE;
+        }
+
+        Object[] newArray = Arrays.copyOf(oldArray, newCapacity);
+        Arrays.fill(newArray, oldCapacity, newArray.length, UNSET);
+        newArray[index] = value;
+        indexedVariables = newArray;
+    }
+
+    Object removeIndexedVariable(int index) {
+        Object[] lookup = indexedVariables;
+        if (index < lookup.length) {
+            Object v = lookup[index];
+            lookup[index] = UNSET;
+            return v;
+        } else {
+            return UNSET;
+        }
+    }
+
+    boolean isIndexedVariableSet(int index) {
+        Object[] lookup = indexedVariables;
+        return index < lookup.length && lookup[index] != UNSET;
+    }
+
+    boolean isCleanerFlagSet(int index) {
+        return cleanerFlags != null && cleanerFlags.get(index);
+    }
+
+    void setCleanerFlag(int index) {
+        if (cleanerFlags == null) {
+            cleanerFlags = new BitSet();
+        }
+        cleanerFlags.set(index);
+    }
+}
